Restructuring and Initial Database Exploration
Directories were added to group python files for better program structure.  Config File and Git Ignore files were added.  Main was restructured to allow for the calling of specific files to run instead of immediately launching the GUI. Basic database interactions were explored.
